Address erratum 8137
David noted in erratum 8137 that no hash algorithm was specified with id-alg-noSignature algorithm in Appendix C. RFC 6402 fixed this by defining sa-noSignature; it defines the hash algorithm as mda-sha1. RFC 6402's ASN.1 module is now in the I-D, so all that is really left is to say what that hash algorithm is in Appendix C. This is the PR for that.